% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits22.R
\name{benefits22}
\alias{benefits22}
\title{Title Calculate Predict2.2 treatment benefits given patient details}
\usage{
benefits22(
  age.start = 57,
  screen = 0,
  size = 29,
  grade = 1,
  nodes = 5,
  er = 1,
  her2 = 1,
  ki67 = 1,
  generation = 3,
  horm = 1,
  traz = 1,
  bis = 1,
  radio = 0,
  delay = 0
)
}
\arguments{
\item{age.start}{Patient age in years}

\item{screen}{Clinically detected = 0, screen detected = 1}

\item{size}{Tumour size mm}

\item{grade}{Tumour grade}

\item{nodes}{Number positive nodes}

\item{er}{ER+ = 1, ER- = 0}

\item{her2}{HER2+ = 1, HER2- = 0, missing = 9}

\item{ki67}{KI67+ = 1, KI67- = 0, missing = 9}

\item{generation}{Chemo generation 0, 2 or 3 only}

\item{horm}{Hormone therapy Yes = 1, no = 0}

\item{traz}{Trastuzumab therapy Yes = 1, no = 0}

\item{bis}{Bisphosphonate therapy Yes = 1, no = 0}

\item{radio}{Radiotherapy Yes = 1, no = 0}

\item{delay}{= 1 Choose extended hormone after 5 years. 0 = before first course}
}
\value{
able of treatment benefits
}
\description{
Title Calculate Predict2.2 treatment benefits given patient details
}
\examples{
benefits22()
}
