% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits222.R
\name{benefits222}
\alias{benefits222}
\title{v2.2 predict disease free survival}
\usage{
benefits222(
  age.start = 57,
  screening = 1,
  size = 20,
  grade = 2,
  nodes = 10,
  er = 1,
  her2 = 1,
  ki67 = 1,
  generation = 1,
  horm = 1,
  traz = 1,
  bis = 1,
  radio = 0,
  delay = 0
)
}
\arguments{
\item{age.start}{Age at time of surgery}

\item{screening}{Clinically detected = 0, screening detected = 1}

\item{size}{Tumour size mm}

\item{nodes}{Number positive nodes}

\item{er}{ER+ = 1, ER- = 0}

\item{her2}{HER2+ = 1, HER2- = 0, missing = 9}

\item{ki67}{KI67+ = 1, KI67- = 0, missing = 9}

\item{generation}{Chemo generation 0, 2 or 3 only}

\item{horm}{Hormone therapy Yes = 1, no = 0}

\item{traz}{Trastuzumab therapy Yes = 1, no = 0}

\item{bis}{Bisphosphonate therapy Yes = 1, no = 0}

\item{radio}{Radiotherapy Yes = 1, no = 0}

\item{delay}{0 or 5 years since surgery for h10 benefit}

\item{tumour}{Tumour grade}
}
\description{
calculation for delay zero, years 1 to 15
}
\examples{
benefits222()
}
