% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits3010.R
\name{benefits3010}
\alias{benefits3010}
\title{Title Calculate benefits of continuing endocrine therapy assuming survival to 5 years}
\usage{
benefits3010()
}
\arguments{
\item{age.start}{Patient age in years}

\item{screen}{Clinically detected = 0, screen detected = 1}

\item{size}{Tumour size mm}

\item{grade}{Tumour grade}

\item{nodes}{Number positive nodes}

\item{er}{ER+ = 1, ER- = 0}

\item{her2}{HER2+ = 1, HER2- = 0, missing = 9}

\item{ki67}{KI67+ = 1, KI67- = 0, missing = 9}

\item{pr}{ptogeserpne satus PR+ = 1 PR- = 0}

\item{generation}{Chemo generation 0, 2 or 3 only}

\item{horm}{Hormone therapy Yes = 1, no = 0}

\item{traz}{Trastuzumab therapy Yes = 1, no = 0}

\item{bis}{Bisphosphonate therapy Yes = 1, no = 0}

\item{radio}{Radiotherapy Yes = 1, no = 0}

\item{heart.gy}{Number of grays radiation to the heart}

\item{smoker}{Never/ex = 0, current = 1}
}
\value{
table of treatment benefits
}
\description{
Title Calculate benefits of continuing endocrine therapy assuming survival to 5 years
}
\examples{
benefits3010()
}
