% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictv3.0-package.R
\docType{package}
\name{Predictv3.0-package}
\alias{Predictv3.0}
\alias{Predictv3.0-package}
\title{Predictv3.0: Publishes Predict v2.2 AND Predictv3.0 as functions}
\description{
Provides the predictv2.2 model and Predictv3.0 as functions.
}
\keyword{internal}
