% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictv30r-package.R
\docType{package}
\name{predictv30r-package}
\alias{predictv30r}
\alias{predictv30r-package}
\title{predictv30r: Publishes Predict v2.2, v3.0, v3.1, v3.2 as functions}
\description{
Provides the predictv2.2 model and Predictv3.0 as functions.
}
\keyword{internal}
