% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREDICT3-package.R
\docType{package}
\name{PREDICT3-package}
\alias{PREDICT3}
\alias{PREDICT3-package}
\title{PREDICT3: Publishes Predict v2.2 AND Predictv3.0 as functions}
\description{
Provides the predictv2.2 model and Predictv3.0 as functions.
}
\keyword{internal}
